refactor: rename and manage db & cache
