+++ controllers/products_controller.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const jieba = require('nodejieba');
 const moment = require('moment');
-const cache = require('../util/redis');
+const cache = require('../util/cache');
 const productModel = require('../models/products_model');
 const userModel = require('../models/users_model');
 const orderModel = require('../models/orders_model');
