+++ package.json
@@ -11,10 +11,9 @@
     "express": "^4.18.2",
     "jest": "^29.3.1",
     "jsonwebtoken": "^8.5.1",
-    "mangodb": "^1.0.0",
     "moment": "^2.29.4",
+    "mongodb": "^4.12.1",
     "mongoose": "^6.7.0",
-    "morgan-body": "^2.6.8",
     "multer": "^1.4.5-lts.1",
     "mysql2": "^2.3.3",
     "neat-csv": "^7.0.0",
