--- util/mysql.js
@@ -1,18 +1,0 @@
-const mysql = require('mysql2/promise');
-
-// my sql database
-require('dotenv').config();
-const { MYSQL_HOST, MYSQL_USER, MYSQL_PASSWORD, MYSQL_DATABASE } = process.env;
-
-const pool = mysql.createPool({
-  host: MYSQL_HOST,
-  user: MYSQL_USER,
-  password: MYSQL_PASSWORD,
-  database: MYSQL_DATABASE,
-  waitForConnections: true,
-  connectionLimit: 10,
-  queueLimit: 10,
-  multipleStatements: true,
-});
-
-module.exports = pool;
